refine Back End - Update Product API

+++ src/main/resources/db/migration/release_1/V1_8__createProduct.sql
@@ -4,7 +4,7 @@ CREATE TABLE PUBLIC.Products (
     productPrice NUMERIC(10, 2) NOT NULL,
     productDescription VARCHAR(250) NOT NULL,
     productTypeId bigint NOT NULL,
-    productCapacity REAL NOT NULL,
+    productCapacity NUMERIC(10, 2) NOT NULL,
     productDownloads bigint NOT NULL DEFAULT 0,
     primary key (productId)
 );
